requirements.txt
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
scikit-learn==0.20.3
pydub==0.23.1
librosa==0.7.0
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
ffmpeg-python==0.2.0
numpy==1.16.2
